规整lua中utils相关测试文件名

+++ llbc/src/core/thread/MessageQueue.cpp
@@ -81,7 +81,7 @@ void LLBC_MessageQueue::Cleanup()
 #if LLBC_TARGET_PLATFORM_NON_WIN32
     _cond.Broadcast();
 #else // LLBC_TARGET_PLATFORM_WIN32
-    _sem.Post(_size);
+    _sem.Post(static_cast<int>(_size));
 #endif // LLBC_TARGET_PLATFORM_NON_WIN32
 
     _size = 0;
